PyMcpServer/pyproject.toml
[project]
name = "mcp-on-cloudrun"
version = "0.1.0"
description = "Example of deploying a MCP server on Cloud Run"
requires-python = ">=3.10"
dependencies = [
    "fastmcp==2.6.1",
    "pyjwt>=2.10.1",
]
